refactor: jwt 응답 글로벌 핸들러 처리
